Added integration test for core functionality
This commit adds an integration test for the core functionality
of the Mimir charm i.e. its ability to setup ingestion of
metrics and alert rules by Mimir. At present the integration
test uses timed sleeps to allow Mimir to ingest metrics
from Grafana agent and raise alerts. In the future this may
be replaced by predicate functions and the block_until API
of Python libjuju.