Add upgrade event handler
This commit adds and upgrade event handler for the Mimir charm.
In response to an upgrade event the Mimir configruation is
regenerated and Mimir restarted. However alert rules are not
aggregated again and reset. This is because if alert rules
change they should cause a relation change event in response
to which the set of all alert rules will be updated.